src/requirements.txt
psycopg2-binary==2.9.9
Flask==3.0.3
requests==2.31.0
Jinja2==3.1.3
redis==5.1.0